requirements.txt
streamlit
transformers
torch
pillow
gtts